Refactor and improve impression detection.
