+++ app/src/main/java/org/wikipedia/feed/news/NewsModule.kt
@@ -3,7 +3,6 @@ package org.wikipedia.feed.news
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -32,7 +31,7 @@ import org.wikipedia.compose.theme.BaseTheme
 import org.wikipedia.compose.theme.WikipediaTheme
 import org.wikipedia.dataclient.WikiSite
 import org.wikipedia.dataclient.page.PageSummary
-import org.wikipedia.feed.CommunityModuleHeader
+import org.wikipedia.feed.CommunityModuleContainer
 import org.wikipedia.feed.noImageCardBackgroundColors
 import org.wikipedia.theme.Theme
 import org.wikipedia.util.StringUtil
@@ -48,23 +47,16 @@ fun NewsModule(
     onHideModuleClick: () -> Unit = {},
     onCardImpression: () -> Unit = {}
 ) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(color = WikipediaTheme.colors.paperColor)
+    val pagerState = rememberPagerState(pageCount = { newsItems.size })
+
+    CommunityModuleContainer(
+        wikiSite = wikiSite,
+        titleResId = R.string.view_card_news_title,
+        subTitleResId = R.string.explore_feed_in_the_news_subtitle,
+        onHideCardClick = onHideCardClick,
+        onHideModuleClick = onHideModuleClick,
+        onCardInView = onCardImpression
     ) {
-
-        val pagerState = rememberPagerState(pageCount = { newsItems.size })
-
-        CommunityModuleHeader(
-            wikiSite = wikiSite,
-            titleResId = R.string.view_card_news_title,
-            subTitleResId = R.string.explore_feed_in_the_news_subtitle,
-            onHideCardClick = onHideCardClick,
-            onHideModuleClick = onHideModuleClick,
-            onCardInView = onCardImpression
-        )
-
         HorizontalPager(
             state = pagerState,
             modifier = Modifier.fillMaxWidth()
